requirements.txt
# AIpanel - Enterprise AI Orchestrator

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0

# Redis (optional - for memory and cache)
redis==5.0.1

# PostgreSQL (optional - for database)
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Security & Encryption
cryptography==41.0.7

# LangChain/LangGraph
langchain==0.1.0
langgraph==0.0.20
langchain-core==0.1.10
langchain-community==0.0.10

# LLM Clients
litellm==1.10.0
openai==1.3.0

# Additional utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1

# Monitoring and logging
python-json-logger==2.0.7
psutil==5.9.6
